requirements.txt
openai==2.7.2
pettingzoo==1.24.3
gymnasium==1.2.3
arch==8.0.0
statsmodels==0.14.4
linearmodels==7.0
scipy==1.15.3
pandas==2.2.3
numpy==2.1.3
wrds==3.5.0
modal==1.4.2
pytest==8.3.4
pytest-timeout==2.4.0
ruptures==1.1.10
seaborn==0.13.2
matplotlib==3.10.0
yfinance==0.2.66
python-dotenv==1.1.0
fredapi==0.5.2
